/network endpoint returns real data

+++ src/main.rs
@@ -11,7 +11,7 @@ fn main() {
     let local_node = consensus::Node::new(config::IP_ADDRESS.to_string(), 
         "0x742d35Cc6634C0532925a3b844Bc454e4438f44e".to_string(), //dummy
         SystemTime::now().duration_since(UNIX_EPOCH)
-            .expect("Time went backwards").as_micros()
+            .expect("Time went backwards").as_micros() as u64
     );
 
     match local_node {
